requirements.txt
streamlit
pandas
numpy
scikit-sklearn